Complete rewrite of UI text to be independent of BMText/Font and to be "marked up".
Added UI text field / area.

TODO: 
* Improve text entry and selection for textarea.
* Improve scroll pane for textarea use.


git-svn-id: http://ardorlabs.svn.cvsdude.com/Ardor3Dv1/trunk@1432 82abaa13-a93f-4eff-8987-b3ae9fc18050

+++ ardor3d-ui/src/main/java/com/ardor3d/extension/ui/util/Alignment.java
@@ -101,7 +101,7 @@ public enum Alignment {
      *            the height of the component or area we are setting position on.
      * @return the new Y position to use.
      */
-    public int alignY(final int availableHeight, final int height) {
-        return (int) (fromBottom() * availableHeight - fromBottom() * height);
+    public double alignY(final double availableHeight, final double height) {
+        return fromBottom() * availableHeight - fromBottom() * height;
     }
 }
